reorganized mock functions
