+++ src/components/bookingForm/index.tsx
@@ -1,6 +1,6 @@
 import { useForm, SubmitHandler, Controller } from 'react-hook-form';
 import AutoCompleteInput from '../autoCompleteInput';
-import { convertToQueryString } from '../services/convert';
+import { convertToQueryString } from '../../services/convert';
 import { SetURLSearchParams } from 'react-router';
 
 type componentProps = {
